[EnhancedNode.getParticipatingNodes] fixed a bug where if two Nodes with same _hash are supplied
- was: the LIFO method used, ie last Node regardless
of the number of its optional properties was used.
It resulted in a bug where if I try to specify
a relationship with Mango.buildSimplifiedDeepEnhancedNode
and try to link to existing Node, a link will be made,
but existing Node will be re-written without its optional
props.
- now: EnhancedNode.getParticipatingNodes picks Node
with most properties, which results in max knowledge
preservation.